[IC-24] feat: 컨테이너 제어부와 모니터링부를 분리 (#17)
