+++ docker/src/main/java/com/initcloud/dockerapi/container/service/ContainerInspectService.java
@@ -1,0 +1,13 @@
+package com.initcloud.dockerapi.container.service;
+
+import java.util.List;
+
+public interface ContainerInspectService<T> {
+
+	T getContainerDetails(String containerId);
+
+	T getContainerMetrics(String containerId);
+
+	List getContainerList();
+
+}
