+++ docker/src/main/java/com/initcloud/dockerapi/container/service/ContainerManageService.java
@@ -1,18 +1,12 @@
 package com.initcloud.dockerapi.container.service;
 
-import java.util.List;
-
-public interface ContainerService<T> {
+public interface ContainerManageService<T> {
 
 	T executeContainer(Integer count);
 
 	T executeContainerForStandBy(Integer count);
 
 	T terminateContainer(String containerId);
-
-	T getContainerDetails(String containerId);
-
-	List getContainerList();
 
 	T pauseContainer(String containerId);
 }
